Refactor mock objects
