+++ tests/bootstrap.php
@@ -21,4 +21,6 @@ spl_autoload_register(function ($class) {
     return;
 });
 
+include 'mock.php';
+
 ?>
